requirements.txt
flask
requests
scikit-learn
python-dotenv
